libmei: fix handler to __u32
this reduce compat layer usage

+++ include/linux/mei_i.h
@@ -78,11 +78,11 @@ struct mei_client_dma_data {
 		unsigned long userptr;
 	};
 	__u32         length;
-	unsigned long handle;
+	__u32         handle;
 };
 
 struct mei_client_dma_handle {
-	unsigned long handle;
+	__u32         handle;
 };
 
 #endif /* _LINUX_MEI_I_H  */
